Add observed state scoping on traffic light agent
Before the scope of observation was set to be every edge intersecting
a traffic light. The issue was leading to some vehicles being counted
twice, for the case of neiboring traffic light intersections.

	EXAMPLE
		| |		| |
	--------|1|-------------| |-------
	     2 3 A	   5 6 7 B
	--------|4|-------------| |------

	SCOPE
	A: 1, 2, 3, 4, 5, 6, 7
	B: 5, 6, 7

Now the scope spams only the half the edge which is closest to the
traffic light, there's no long an overlap of scopes. Additionaly an
environement variable "filter_incoming_edges" was added to further
limit scopes to account only for inflows (True), only outflows (False)
or both (None).

	EXAMPLE
		| |		| |
	--------|1|-------------| |-------
	     2 3 A	   5 6 7 B
	--------|4|-------------| |------

	SCOPE
	A: 1, 2, 3, 4
	B: 5, 6, 7