Parse emissions.csv file coming from SUMO's repository (#5)
ISSUE:
------

When running long simulations, the environment was unable to convert emission.xml file
to emission.csv. The solution was using a proprietary -- SUMO's script to make the
so called conversion. However the resulting emissions.csv would differ from the generated
from flow. This commit fixes this issue and should parse both Flow's generated
emissions.csv, as well was SUMO's.