Fix a network connection bug on grid       (#13)
ISSUE
-----

A car generated on a vertical edge with a single lane, turned to the
left on the inner most lane and stopped on the red light and did not
re-accelerate. Traffic was accumulating on the lane preventing other
other vehicles to occupy the lane causing gridlocks.

FIX:
---

The problem was related to the lane changing policy. The
grid.net.xml was edited to allow for inbound vehicles to occupy
both lanes (inner most and outer most). As of now vehicles do not
change lanes once they occupy an edge.