Reduce train, evaluate and re-train cycle  (#14)
ISSUE:
------

Reduce the hyper parameter search cycle

FIX:
---

The longest task during training was the generation of the emissions
file which always grows do the decimal of the GBs and took a long time
to generate. The shift in scripts was always allow the info json (for
aggregate performance metrics) and the pickle (for recovering the
training models) to be generated by default. While enabling emissions'
file generation only on evaluation.