Split phase on observation space            (#1)
ISSUE:
-----

Observation states were defined by the incoming and outgoing edges
wrt the intersection -- resulting in an agent being unable to
separate a north-south flow from east-west flow.

FIX:
----
Split spaces not only by incoming and outgoing edges but by the lanes which are controlled by certain phases. Although more robust tests are needed to assertain wheter or not the agent is able to generate consistent policies the preliminary results seem to be promissing.